fix: flacky tests post items
